feat(endpoint-media): remove frontend interface
