+++ packages/endpoint-media/index.js
@@ -1,7 +1,6 @@
 import express from "express";
 import multer from "multer";
 import { mediaController } from "./lib/controllers/media.js";
-import { micropubController } from "./lib/controllers/micropub.js";
 
 const defaults = {
   mountPath: "/media",
@@ -17,28 +16,14 @@ export const MediaEndpoint = class {
     this._router = express.Router(); // eslint-disable-line new-cap
   }
 
-  navigationItems(application) {
-    if (application.hasDatabase) {
-      return {
-        href: this.options.mountPath,
-        text: "media.title",
-      };
-    }
-  }
-
   get routes() {
     const router = this._router;
     const multipartParser = multer({
       storage: multer.memoryStorage(),
     });
 
-    // Application
-    router.get("/", mediaController.files);
-    router.get("/:id", mediaController.file);
-
-    // Micropub API
-    router.get("/", micropubController.query);
-    router.post("/", multipartParser.single("file"), micropubController.upload);
+    router.get("/", mediaController.query);
+    router.post("/", multipartParser.single("file"), mediaController.upload);
 
     return router;
   }
