--- packages/endpoint-media/locales/de.json
@@ -1,11 +1,0 @@
-{
-  "media": {
-    "download": "Download-Datei",
-    "files": {
-      "none": "Keine Dateien",
-      "title": "Hochgeladene Dateien"
-    },
-    "properties": "Eigenschaften",
-    "title": "Dateien"
-  }
-}
