--- packages/endpoint-media/locales/en.json
@@ -1,11 +1,0 @@
-{
-  "media": {
-    "title": "Files",
-    "download": "Download file",
-    "properties": "Properties",
-    "files": {
-      "title": "Uploaded files",
-      "none": "No files"
-    }
-  }
-}
