--- packages/endpoint-media/locales/es.json
@@ -1,11 +1,0 @@
-{
-  "media": {
-    "download": "Descargar archivo",
-    "files": {
-      "none": "No hay archivos",
-      "title": "Archivos cargados"
-    },
-    "properties": "Propiedades",
-    "title": "Archivos"
-  }
-}
