--- packages/endpoint-media/locales/fr.json
@@ -1,11 +1,0 @@
-{
-  "media": {
-    "download": "Télécharger le fichier",
-    "files": {
-      "none": "Aucun fichier",
-      "title": "Fichiers téléchargés"
-    },
-    "properties": "Propriétés",
-    "title": "Fichiers"
-  }
-}
