--- packages/endpoint-media/locales/id.json
@@ -1,11 +1,0 @@
-{
-  "media": {
-    "download": "Unduh file",
-    "files": {
-      "none": "Tidak ada file",
-      "title": "File yang diunggah"
-    },
-    "properties": "Properti",
-    "title": "File"
-  }
-}
