--- packages/endpoint-media/locales/nl.json
@@ -1,11 +1,0 @@
-{
-  "media": {
-    "download": "Bestand downloaden",
-    "files": {
-      "none": "Geen bestanden",
-      "title": "Geüploade bestanden"
-    },
-    "properties": "Eigenschappen",
-    "title": "Bestanden"
-  }
-}
