--- packages/endpoint-media/locales/pt.json
@@ -1,11 +1,0 @@
-{
-  "media": {
-    "download": "Transferir ficheiro",
-    "files": {
-      "none": "Sem ficheiros",
-      "title": "Ficheiros carregados"
-    },
-    "properties": "Propriedades",
-    "title": "Ficheiros"
-  }
-}
