--- packages/endpoint-media/views/file.njk
@@ -1,20 +1,0 @@
-{% extends "document.njk" %}
-
-{% set title = file.filename %}
-
-{% block content %}
-  {% set type = file["post-type"] %}
-  {% if type == "audio" %}
-    <audio src="{{ file.url }}" controls>
-  {% elif type == "photo" %}
-    <img src="{{ file.url }}" onerror="this.src='/assets/not-found.svg'; this.className='not-found';">
-  {% elif type == "video" %}
-    <video src="{{ file.url }}" controls>
-  {% else %}
-    <a href="{{ file.url }}" download>{{ __("media.download") }}</a>
-  {% endif %}
-  {{ summary({
-    title: __("media.properties"),
-    rows: file | summaryRows
-  }) }}
-{% endblock %}
