--- packages/endpoint-media/views/files.njk
@@ -1,12 +1,0 @@
-{% extends "document.njk" %}
-
-{% block content %}
-  {%- if files.length > 0 %}
-  {{ fileGrid({
-    items: files
-  }) }}
-  {{ pagination(pages(page, limit, count)) }}
-  {%- else -%}
-  <p>{{ __("media.files.none") }}<p>
-  {%- endif %}
-{% endblock %}
